cdk/requirements.txt
aws-cdk-lib>=2.0.0
constructs>=10.0.0
cdk-docker-image-deployment>=0.0.483

